src/main/java/com/epam/edu/spring/core/template/service/SimpleItemService.java
package com.epam.edu.spring.core.template.service;

import com.epam.edu.spring.core.template.entity.Item;
import com.epam.edu.spring.core.template.repository.ItemRepository;
import com.epam.edu.spring.core.template.validator.ItemValidator;
import org.springframework.beans.factory.annotation.Autowired;

public class SimpleItemService implements ItemService {

    @Autowired
    private ItemRepository itemRepository;
    @Autowired
    private ItemValidator itemValidator;

    @Override
    public Item getById(long id) {
        return itemRepository.getById(id);
    }

    @Override
    public boolean createItem(Item item) {
        if (!itemValidator.isItemValid(item)) {
            return false;
        }
        return itemRepository.createItem(item);
    }

    public SimpleItemService(ItemRepository itemRepository, ItemValidator itemValidator) {
        this.itemRepository = itemRepository;
        this.itemValidator = itemValidator;
    }

    public SimpleItemService() {

    }

    public void setItemRepository(ItemRepository itemRepository) {
        this.itemRepository = itemRepository;
    }

    public void setItemValidator(ItemValidator itemValidator) {
        this.itemValidator = itemValidator;
    }
}
